+++ berserk/types/core/aliases.py
@@ -1,7 +1,0 @@
-"""Aliases for common types"""
-from __future__ import annotations
-
-from typing_extensions import TypeAlias
-
-ID: TypeAlias = str
-Username: TypeAlias = str