+++ CHANGELOG.rst
@@ -1,14 +1,24 @@
 Changelog
 =========
 
+Next release
+------------
+
+* Added ``client.bulk_pairings`` to create and manage bulk pairings
+  * ``client.bulk_pairings.get_upcoming`` to get upcoming bulk pairings you created
+  * ``client.bulk_pairings.create`` to create a bulk pairing
+  * ``client.bulk_pairings.start_clocks`` to start the clocks of a bulk pairing
+  * ``client.bulk_pairings.cancel`` to cancel a bulk pairing
+* Added better return types for ``client.account.get`` and ``client.account.get_preferences``
+
+Thanks to @Virinas-code for their contributions to this release.
+
 v0.12.8 (2023-07-25)
 --------------------
 
 * Added ``client.opening_explorer.get_lichess_games`` to request opening explorer data for a given position based on Lichess games
-* Added ``client.bulk_pairings`` to create and manage bulk pairings
-* Reworked ``client.account`` with better typing
 
-Thanks to @rpesche and @Virinas-code for their contributions to this release.
+Thanks to @rpesche for their contributions to this release.
 
 v0.12.7 (2023-07-15)
 --------------------